🚀 categorizing request classes
