+++ backend/abba/src/main/java/com/zon/abba/member/controller/RecommendController.java
@@ -1,8 +1,8 @@
 package com.zon.abba.member.controller;
 
 import com.zon.abba.common.response.ResponseBody;
-import com.zon.abba.member.request.EmailRequest;
-import com.zon.abba.member.request.ListRecommendRequest;
+import com.zon.abba.member.request.email.EmailRequest;
+import com.zon.abba.member.request.recommend.ListRecommendRequest;
 import com.zon.abba.member.service.RecommendService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
