+++ backend/abba/src/main/java/com/zon/abba/member/request/member/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.zon.abba.member.request;
+package com.zon.abba.member.request.member;
 
 import lombok.*;
