+++ backend/abba/src/main/java/com/zon/abba/member/request/member/MemberListRequest.java
@@ -1,4 +1,4 @@
-package com.zon.abba.member.request;
+package com.zon.abba.member.request.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,11 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ListRecommendRequest {
+public class MemberListRequest {
     private Integer pageNo;
     private Integer pageSize;
+    private String sort;
+    private String sortValue;
+    private String filter;
+    private String filterValue;
 }
