+++ backend/abba/src/main/java/com/zon/abba/member/request/member/MemberRoleRequest.java
@@ -1,4 +1,4 @@
-package com.zon.abba.member.request;
+package com.zon.abba.member.request.member;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -10,10 +10,9 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class MemberPasswordRequest {
-    @JsonProperty("email")
-    private String email;
-
-    @JsonProperty("password")
-    private String password;
+public class MemberRoleRequest {
+    @JsonProperty("memberID")
+    private String memberID;
+    @JsonProperty("role")
+    private String role;
 }
