+++ backend/abba/src/main/java/com/zon/abba/member/request/oauth/GoogleAccessTokenRequest.java
@@ -1,4 +1,4 @@
-package com.zon.abba.member.request;
+package com.zon.abba.member.request.oauth;
 
 import lombok.*;
